requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
certifi==2024.7.4
click==8.1.7
exceptiongroup==1.2.2
fakeredis==2.23.5
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
